backend/requirements.txt
Django==5.2.5
djangorestframework==3.14.0
django-cors-headers==4.3.1
psutil==5.9.6
requests==2.31.0
python-dotenv==1.0.0
